Move SNBGraph into grakn-test-tools
